Move frontend tests into tests/frontend/
Group the frontend (server/normalizer) tests under tests/frontend/,
leaving the loader tests at tests/ for a future tests/loader/ move.
Relative imports (..helpers) and resources paths (parent.parent)
updated for the new depth; shared helpers/, resources/, and conftest.py
stay at tests/.
